refactor(): move CorsProcessor to CorsHeaderFilterAdapter
